Fix partial digests for SHA1 and SHA256.
The l2n() macro assumes it is working on a char * pointer as it advances
the pointer as it reads.  When 'temp' was a pointer to an unsigned int,
the bytes of the partial hashes were scattered at a 4 byte stride
(and overflowed the md).

While here, simplify the SHA384 case by using l2n8() so it is portable
and doesn't require OS-specific macros like htonll() or htobe64().